Move TM fetch,handl,poller out of common
The fetcher, handler, and poller were put in common because they
were originally intended to be generic, usable by other apps. But
they just aren't generic enough to warrant a generic location. The
big hindrance is that Go typing requires static objects, and has no
paradigm for generic funcs or types. If another app needs similar
functionality in the future, we can revisit genericising these.

Also removes the unused InsomniacPoll.
